refactor: 💡 ExoPlayer HLSMediaSource (Segment sequence)

+++ android/src/main/java/com/guichaguri/trackplayer/service/MusicBinder.java
@@ -42,7 +42,7 @@ public class MusicBinder extends Binder {
     public void updateOptions(Bundle bundle) {
         manager.setStopWithApp(bundle.getBoolean("stopWithApp", false));
         manager.setAlwaysPauseOnInterruption(bundle.getBoolean("alwaysPauseOnInterruption", false));
-        manager.setDuckingVolumeMultiplier(bundle.getFloat("duckingVolumeMultiplier", 0.5F));
+        manager.setDuckingVolumeMultiplier((float)bundle.getDouble("duckingVolumeMultiplier", 0.5));
         manager.getMetadata().updateOptions(bundle);
     }
